+++ utils/mocks/stake_manager_utils.go
@@ -170,14 +170,14 @@ func (_m *StakeManagerUtils) MinSafeRazor(client *ethclient.Client) (*big.Int, e
 }
 
 // WithdrawInitiationPeriod provides a mock function with given fields: client
-func (_m *StakeManagerUtils) WithdrawInitiationPeriod(client *ethclient.Client) (uint8, error) {
+func (_m *StakeManagerUtils) WithdrawInitiationPeriod(client *ethclient.Client) (uint16, error) {
 	ret := _m.Called(client)
 
-	var r0 uint8
-	if rf, ok := ret.Get(0).(func(*ethclient.Client) uint8); ok {
+	var r0 uint16
+	if rf, ok := ret.Get(0).(func(*ethclient.Client) uint16); ok {
 		r0 = rf(client)
 	} else {
-		r0 = ret.Get(0).(uint8)
+		r0 = ret.Get(0).(uint16)
 	}
 
 	var r1 error
@@ -191,14 +191,14 @@ func (_m *StakeManagerUtils) WithdrawInitiationPeriod(client *ethclient.Client) 
 }
 
 // WithdrawLockPeriod provides a mock function with given fields: client
-func (_m *StakeManagerUtils) WithdrawLockPeriod(client *ethclient.Client) (uint8, error) {
+func (_m *StakeManagerUtils) WithdrawLockPeriod(client *ethclient.Client) (uint16, error) {
 	ret := _m.Called(client)
 
-	var r0 uint8
-	if rf, ok := ret.Get(0).(func(*ethclient.Client) uint8); ok {
+	var r0 uint16
+	if rf, ok := ret.Get(0).(func(*ethclient.Client) uint16); ok {
 		r0 = rf(client)
 	} else {
-		r0 = ret.Get(0).(uint8)
+		r0 = ret.Get(0).(uint16)
 	}
 
 	var r1 error
