+++ utils/struct-utils.go
@@ -57,12 +57,12 @@ func (b BlockManagerStruct) GetBlockIndexToBeConfirmed(client *ethclient.Client)
 	return blockManager.BlockIndexToBeConfirmed(&opts)
 }
 
-func (s StakeManagerStruct) WithdrawInitiationPeriod(client *ethclient.Client) (uint8, error) {
+func (s StakeManagerStruct) WithdrawInitiationPeriod(client *ethclient.Client) (uint16, error) {
 	stakeManager, opts := UtilsInterface.GetStakeManagerWithOpts(client)
 	return stakeManager.WithdrawInitiationPeriod(&opts)
 }
 
-func (s StakeManagerStruct) WithdrawLockPeriod(client *ethclient.Client) (uint8, error) {
+func (s StakeManagerStruct) WithdrawLockPeriod(client *ethclient.Client) (uint16, error) {
 	stakeManager, opts := UtilsInterface.GetStakeManagerWithOpts(client)
 	return stakeManager.WithdrawLockPeriod(&opts)
 }
